requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.31.1
python-socketio==5.11.4
python-engineio==4.10.1
motor==3.6.0
requests==2.32.3
PyJWT==2.9.0
python-multipart==0.0.6
aiofiles==23.2.1
opencv-python==4.8.1.78
google-generativeai==0.3.2
PyPDF2==3.0.1
gTTS==2.4.0
pydub==0.25.1
pydantic==2.5.0
